requirements.txt
python-telegram-bot==13.5
beautifulsoup4==4.9.3
requests==2.25.1